+++ client/src/components/board/BoardSearchBar.js
@@ -3,6 +3,7 @@ import AppContext from "context/AppContext";
 import BoardContext from "context/BoardContext";
 import React, {useContext} from 'react';
 import {TextareaAutosize} from "react-autosize-textarea/lib/TextareaAutosize";
+import tinycolor from "tinycolor2";
 import {UiBadge} from "ui";
 import {UiDropdownElement, UiSelectableDropdown} from "ui/dropdown";
 import {UiFormControl} from "ui/form";
@@ -22,7 +23,7 @@ const SearchBar = styled(TextareaAutosize)`
 
 const BoardSearchBar = ({searchQuery, setSearchQuery}) => {
     const {user, onLocalPreferencesUpdate} = useContext(AppContext);
-    const {allIdeas, openedIdeas, closedIdeas} = useContext(BoardContext).data;
+    const {tags, allIdeas, openedIdeas, closedIdeas} = useContext(BoardContext).data;
     const queryRef = React.useRef();
     if (searchQuery === "" && queryRef.current) {
         queryRef.current.value = "";
@@ -32,6 +33,7 @@ const BoardSearchBar = ({searchQuery, setSearchQuery}) => {
         {closed: {name: "Closed", data: closedIdeas}},
         {all: {name: "All", data: allIdeas}}
     ];
+    tags.forEach(tag => filters.push({["tag:" + tag.id]: {name: <UiBadge color={tinycolor(tag.color)}>{tag.name}</UiBadge>, data: null}}));
     const sorts = [
         {trending: "Trending"},
         {voters_desc: "Most Voted"},
@@ -42,12 +44,12 @@ const BoardSearchBar = ({searchQuery, setSearchQuery}) => {
     let searchTimeout;
     const filterCurrentValue = Object.values(filters.find(obj => {
         return Object.keys(obj)[0] === (user.localPreferences.ideas.filter || "opened")
-    }))[0].name;
+    }) || filters[0])[0].name;
     const filterValues = filters.map(val => {
         const key = Object.keys(val)[0];
         const value = Object.values(val)[0];
         return <UiDropdownElement key={key} onClick={() => onLocalPreferencesUpdate({...user.localPreferences, ideas: {...user.localPreferences.ideas, filter: key}})}>
-            <React.Fragment>{value.name} <UiBadge className={"float-right pt-1"}>{value.data}</UiBadge></React.Fragment>
+            <React.Fragment>{value.name} {value.data && <UiBadge className={"float-right pt-1"}>{value.data}</UiBadge>}</React.Fragment>
         </UiDropdownElement>
     });
     const sortCurrentValue = Object.values(sorts.find(obj => {
@@ -69,7 +71,7 @@ const BoardSearchBar = ({searchQuery, setSearchQuery}) => {
             <UiFormControl label={"Search Ideas"} as={SearchBar} innerRef={queryRef} maxLength={40} rows={1} maxRows={1} defaultValue={searchQuery} placeholder={"Search"} onInput={() => {
                 clearTimeout(searchTimeout);
                 searchTimeout = setTimeout(() => setSearchQuery(queryRef.current.value.substring(0, 40)), 500);
-            }}  aria-label={"Search bar"}/>
+            }} aria-label={"Search bar"}/>
         </UiCol>
     </React.Fragment>
 };