+++ client/src/ui/UiBadge.js
@@ -1,4 +1,5 @@
 import styled from "@emotion/styled";
+import {WCAG_AA_CONTRAST} from "App";
 import AppContext from "context/AppContext";
 import PropTypes from "prop-types";
 import React, {useContext} from "react";
@@ -26,12 +27,16 @@ const UiBadge = (props) => {
     if (context.user.darkMode) {
         badgeColor = badgeColor.lighten(10);
         //if still not readable, increase again
-        if (tinycolor.readability(badgeColor, "#282828") < 2.5) {
+        if (tinycolor.readability(badgeColor, "#282828") < WCAG_AA_CONTRAST) {
             badgeColor = badgeColor.lighten(25);
         }
         return <Badge theme={badgeColor.clone().setAlpha(.1).toString()} style={{color: badgeColor, fontWeight: "bold", ...style}} ref={innerRef} {...otherProps}>{children}</Badge>
+    } else {
+        if (tinycolor.readability(badgeColor, "#fff") < WCAG_AA_CONTRAST) {
+            badgeColor = badgeColor.darken(10);
+        }
     }
-    return <Badge theme={badgeColor.toString()} ref={innerRef} style={style} {...otherProps}>{children}</Badge>
+    return <Badge theme={badgeColor.toString()} style={{fontWeight: "500", color: badgeColor.toString(), backgroundColor: badgeColor.setAlpha(.1).toString(), ...style}} ref={innerRef} {...otherProps}>{children}</Badge>
 };
 
 export {UiBadge};